Separated components out for easier access to each individual piece on the Comparison page.
