+++ src/components/pages/ComparisonPage/Container.js
@@ -1,0 +1,15 @@
+import React, { useState } from 'react';
+import styled from 'styled-components';
+import { DescriptionSection } from './DecriptionSection';
+import { Header } from './Header';
+import { MainVisual } from './MainVisual';
+
+export const Container = () => {
+  return (
+    <>
+      <Header />
+      <MainVisual />
+      <DescriptionSection />
+    </>
+  );
+};
